:recycle: tweak image extraction module
