+++ src/imageOperations/multiReceiptsExtractor/multiReceiptsExtractor.ts
@@ -1,11 +1,11 @@
 import { PDFDocument, PDFImage, PDFPage } from "pdf-lib";
-import { MindeeError, MindeeMimeTypeError } from "../errors";
-import { Polygon } from "../geometry";
-import { MultiReceiptsDetectorV1 } from "../product";
+import { MindeeError, MindeeMimeTypeError } from "../../errors";
+import { Polygon } from "../../geometry";
+import { MultiReceiptsDetectorV1 } from "../../product";
 import { ExtractedMultiReceiptImage } from "./extractedMultiReceiptImage";
-import { LocalInputSource } from "../input/base";
-import { extractFromPage } from "./imageExtractor";
-import { PositionField } from "../parsing/standard";
+import { LocalInputSource } from "../../input/base";
+import { extractFromPage } from "../common/imageExtractor";
+import { PositionField } from "../../parsing/standard";
 
 /**
  * Given a page and a set of coordinates, extracts & assigns individual receipts to an ExtractedMultiReceiptImage
